# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SEAPEDIA API",
    "version": "1.0.0",
    "description": "Multi-role e-commerce platform API. Supports Admin, Seller, Buyer, and Driver workflows. Protected endpoints require a Bearer JWT token in the Authorization header."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8787",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/stores/{slug}": {
      "get": {
        "tags": [
          "Stores"
        ],
        "summary": "Get store by slug (public)",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Store with products and stats"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/stores/me": {
      "get": {
        "tags": [
          "Stores - Seller"
        ],
        "summary": "Get seller's own store",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Seller store data"
          }
        }
      },
      "post": {
        "tags": [
          "Stores - Seller"
        ],
        "summary": "Create a store",
        "description": "Store name must be globally unique.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "logoUrl": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Store created"
          },
          "400": {
            "description": "Name already taken"
          }
        }
      },
      "put": {
        "tags": [
          "Stores - Seller"
        ],
        "summary": "Update store",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Store updated"
          }
        }
      }
    },
    "/wallet": {
      "get": {
        "tags": [
          "Wallet"
        ],
        "summary": "Get wallet balance",
        "description": "Works for Buyer, Seller, and Driver roles.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Wallet with current balance"
          }
        }
      }
    },
    "/wallet/topup": {
      "post": {
        "tags": [
          "Wallet"
        ],
        "summary": "Top up wallet (dummy simulation)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "amount"
                ],
                "properties": {
                  "amount": {
                    "type": "number",
                    "minimum": 1000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Top up successful"
          }
        }
      }
    },
    "/wallet/transactions": {
      "get": {
        "tags": [
          "Wallet"
        ],
        "summary": "Wallet transaction history",
        "description": "Returns all mutations: TOP_UP, PAYMENT, INCOME, REFUND.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Array of wallet mutations"
          }
        }
      }
    },
...
    "/admin/stores": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List all stores (admin)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "All stores with owner and product count"
          }
        }
      }
    },
...
  }
}